# openapi/workledger.openapi.json
{
...
  "info": { "title": "WorkLedger API", "version": "1.0.0" },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/me/csrf": {
      "get": {
        "description": "Bootstraps the in-memory session-bound CSRF token used by WorkLedger unsafe requests.",
        "operationId": "getSelfCsrfToken",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "additionalProperties": false,
                      "properties": {
                        "token": { "maxLength": 256, "minLength": 32, "type": "string" }
                      },
                      "required": ["token"],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": {
                        "idempotentReplay": { "type": "boolean" },
                        "requestId": {
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                          "type": "string"
                        }
                      },
                      "required": ["requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["data", "meta"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "enum": [
                            "ABSENCE_ALREADY_DECIDED",
                            "ABSENCE_AMBIGUOUS_LOCAL_TIME",
                            "ABSENCE_CANNOT_CANCEL",
                            "ABSENCE_COVERAGE_INVALID",
                            "ABSENCE_DURATION_NOT_ALLOWED",
                            "ABSENCE_INSUFFICIENT_BALANCE",
                            "ABSENCE_NONEXISTENT_LOCAL_TIME",
                            "ABSENCE_OVERLAP",
                            "ABSENCE_OVERRIDE_REASON_REQUIRED",
                            "ABSENCE_POLICY_INACTIVE",
                            "ABSENCE_REPORT_CANNOT_REJECT",
                            "ABSENCE_REQUEST_NOTE_NOT_ALLOWED",
                            "ABSENCE_RETROACTIVE_LIMIT",
                            "ABSENCE_STATE_CHANGED",
                            "ACCESS_DENIED",
                            "APPROVAL_OUT_OF_SCOPE",
                            "APPROVAL_REASON_REQUIRED",
                            "APPROVAL_SELF_NOT_ALLOWED",
                            "APPROVAL_STATE_CONFLICT",
                            "ATTENDANCE_ALREADY_OFF_WORK",
                            "ATTENDANCE_ALREADY_ON_BREAK",
                            "ATTENDANCE_ALREADY_WORKING",
                            "ATTENDANCE_AMBIGUOUS_LOCAL_TIME",
                            "ATTENDANCE_BREAK_CONFIRMATION_REQUIRED",
                            "ATTENDANCE_FUTURE_EVENT",
                            "ATTENDANCE_INCOMPLETE",
                            "ATTENDANCE_INVALID_EVENT_ORDER",
                            "ATTENDANCE_INVALID_EVENT_PRECISION",
                            "ATTENDANCE_NONEXISTENT_LOCAL_TIME",
                            "ATTENDANCE_NOT_ON_BREAK",
                            "ATTENDANCE_NOT_WORKING",
                            "ATTENDANCE_OVERLAP",
                            "ATTENDANCE_STATE_CHANGED",
                            "AUTH_CSRF_INVALID",
                            "AUTH_INVALID_CREDENTIALS",
                            "AUTH_INVITATION_INVALID_OR_EXPIRED",
                            "AUTH_ORIGIN_INVALID",
                            "AUTH_REQUIRED",
                            "AUTH_RESET_INVALID_OR_EXPIRED",
                            "AUTH_SESSION_EXPIRED",
                            "AUTH_SESSION_NOT_FRESH",
                            "CORRECTION_UNRESOLVED",
                            "DATABASE_UNAVAILABLE",
                            "EMPLOYEE_NOT_FOUND",
                            "IDEMPOTENCY_KEY_CONFLICT",
                            "IDEMPOTENCY_KEY_INVALID",
                            "IDEMPOTENCY_KEY_REQUIRED",
                            "INTERNAL_ERROR",
                            "LEDGER_SOURCE_MISMATCH",
                            "MALFORMED_REQUEST",
                            "ORGANIZATION_TIMEZONE_LOCKED",
                            "PERIOD_ADJUSTMENT_REQUIRED",
                            "PERIOD_ALREADY_SUBMITTED",
                            "PERIOD_LEDGER_MISMATCH",
                            "PERIOD_LOCKED",
                            "PERIOD_NOT_READY",
                            "PERIOD_REOPEN_REQUIRED",
                            "PERIOD_SOURCE_CHANGED",
                            "PERIOD_STATE_CONFLICT",
                            "PERIOD_VERSION_CONFLICT",
                            "PERIOD_WARNING_ACKNOWLEDGEMENT_REQUIRED",
                            "POLICY_CONFIGURATION_INVALID",
                            "POLICY_NOT_ASSIGNED",
                            "RATE_LIMITED",
                            "RECORD_VERSION_CONFLICT",
                            "REQUEST_TOO_LARGE",
                            "ROUTE_NOT_FOUND",
                            "SCHEDULE_ASSIGNMENT_OVERLAP",
                            "SCHEDULE_NOT_ASSIGNED",
                            "TIME_ACCOUNT_LEDGER_DUPLICATE_ENTRY",
                            "TIME_ACCOUNT_LEDGER_DUPLICATE_SOURCE",
                            "TIME_ACCOUNT_LEDGER_SCOPE_MISMATCH",
                            "TIME_ACCOUNT_LEDGER_TOTAL_INVALID",
                            "UNSUPPORTED_MEDIA_TYPE",
                            "VALIDATION_FAILED"
                          ],
                          "type": "string"
                        },
                        "context": {
                          "additionalProperties": {
                            "anyOf": [
                              {
                                "anyOf": [
                                  { "maxLength": 256, "type": "string" },
                                  {
                                    "maximum": 9007199254740991,
                                    "minimum": -9007199254740991,
                                    "type": "integer"
                                  },
                                  { "type": "boolean" },
                                  { "type": "null" }
                                ]
                              },
                              {
                                "items": {
                                  "anyOf": [
                                    { "maxLength": 256, "type": "string" },
                                    {
                                      "maximum": 9007199254740991,
                                      "minimum": -9007199254740991,
                                      "type": "integer"
                                    },
                                    { "type": "boolean" },
                                    { "type": "null" }
                                  ]
                                },
                                "maxItems": 50,
                                "type": "array"
                              }
                            ]
                          },
                          "propertyNames": {
                            "pattern": "^[A-Za-z][A-Za-z0-9]{0,63}$",
                            "type": "string"
                          },
                          "type": "object"
                        },
                        "fields": {
                          "additionalProperties": {
                            "items": {
                              "additionalProperties": false,
                              "properties": {
                                "code": {
                                  "enum": [
                                    "INVALID_FORMAT",
                                    "INVALID_TYPE",
                                    "INVALID_VALUE",
                                    "REQUIRED",
                                    "UNKNOWN_FIELD",
                                    "VALUE_TOO_LARGE",
                                    "VALUE_TOO_SMALL"
                                  ],
                                  "type": "string"
                                },
                                "message": { "maxLength": 256, "minLength": 1, "type": "string" }
                              },
                              "required": ["code", "message"],
                              "type": "object"
                            },
                            "maxItems": 10,
                            "minItems": 1,
                            "type": "array"
                          },
                          "propertyNames": { "maxLength": 256, "minLength": 1, "type": "string" },
                          "type": "object"
                        },
                        "message": { "maxLength": 256, "minLength": 1, "type": "string" },
                        "requestId": {
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                          "type": "string"
                        }
                      },
                      "required": ["code", "message", "requestId"],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": { "idempotentReplay": { "type": "boolean" } },
                      "required": ["idempotentReplay"],
                      "type": "object"
                    }
                  },
                  "required": ["error"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "enum": [
                            "ABSENCE_ALREADY_DECIDED",
                            "ABSENCE_AMBIGUOUS_LOCAL_TIME",
                            "ABSENCE_CANNOT_CANCEL",
                            "ABSENCE_COVERAGE_INVALID",
                            "ABSENCE_DURATION_NOT_ALLOWED",
                            "ABSENCE_INSUFFICIENT_BALANCE",
                            "ABSENCE_NONEXISTENT_LOCAL_TIME",
                            "ABSENCE_OVERLAP",
                            "ABSENCE_OVERRIDE_REASON_REQUIRED",
                            "ABSENCE_POLICY_INACTIVE",
                            "ABSENCE_REPORT_CANNOT_REJECT",
                            "ABSENCE_REQUEST_NOTE_NOT_ALLOWED",
                            "ABSENCE_RETROACTIVE_LIMIT",
                            "ABSENCE_STATE_CHANGED",
                            "ACCESS_DENIED",
                            "APPROVAL_OUT_OF_SCOPE",
                            "APPROVAL_REASON_REQUIRED",
                            "APPROVAL_SELF_NOT_ALLOWED",
                            "APPROVAL_STATE_CONFLICT",
                            "ATTENDANCE_ALREADY_OFF_WORK",
                            "ATTENDANCE_ALREADY_ON_BREAK",
                            "ATTENDANCE_ALREADY_WORKING",
                            "ATTENDANCE_AMBIGUOUS_LOCAL_TIME",
                            "ATTENDANCE_BREAK_CONFIRMATION_REQUIRED",
                            "ATTENDANCE_FUTURE_EVENT",
                            "ATTENDANCE_INCOMPLETE",
                            "ATTENDANCE_INVALID_EVENT_ORDER",
                            "ATTENDANCE_INVALID_EVENT_PRECISION",
                            "ATTENDANCE_NONEXISTENT_LOCAL_TIME",
                            "ATTENDANCE_NOT_ON_BREAK",
                            "ATTENDANCE_NOT_WORKING",
                            "ATTENDANCE_OVERLAP",
                            "ATTENDANCE_STATE_CHANGED",
                            "AUTH_CSRF_INVALID",
                            "AUTH_INVALID_CREDENTIALS",
                            "AUTH_INVITATION_INVALID_OR_EXPIRED",
                            "AUTH_ORIGIN_INVALID",
                            "AUTH_REQUIRED",
                            "AUTH_RESET_INVALID_OR_EXPIRED",
                            "AUTH_SESSION_EXPIRED",
                            "AUTH_SESSION_NOT_FRESH",
                            "CORRECTION_UNRESOLVED",
                            "DATABASE_UNAVAILABLE",
                            "EMPLOYEE_NOT_FOUND",
                            "IDEMPOTENCY_KEY_CONFLICT",
                            "IDEMPOTENCY_KEY_INVALID",
                            "IDEMPOTENCY_KEY_REQUIRED",
                            "INTERNAL_ERROR",
                            "LEDGER_SOURCE_MISMATCH",
                            "MALFORMED_REQUEST",
                            "ORGANIZATION_TIMEZONE_LOCKED",
                            "PERIOD_ADJUSTMENT_REQUIRED",
                            "PERIOD_ALREADY_SUBMITTED",
                            "PERIOD_LEDGER_MISMATCH",
                            "PERIOD_LOCKED",
                            "PERIOD_NOT_READY",
                            "PERIOD_REOPEN_REQUIRED",
                            "PERIOD_SOURCE_CHANGED",
                            "PERIOD_STATE_CONFLICT",
                            "PERIOD_VERSION_CONFLICT",
                            "PERIOD_WARNING_ACKNOWLEDGEMENT_REQUIRED",
                            "POLICY_CONFIGURATION_INVALID",
                            "POLICY_NOT_ASSIGNED",
                            "RATE_LIMITED",
                            "RECORD_VERSION_CONFLICT",
                            "REQUEST_TOO_LARGE",
                            "ROUTE_NOT_FOUND",
                            "SCHEDULE_ASSIGNMENT_OVERLAP",
                            "SCHEDULE_NOT_ASSIGNED",
                            "TIME_ACCOUNT_LEDGER_DUPLICATE_ENTRY",
                            "TIME_ACCOUNT_LEDGER_DUPLICATE_SOURCE",
                            "TIME_ACCOUNT_LEDGER_SCOPE_MISMATCH",
                            "TIME_ACCOUNT_LEDGER_TOTAL_INVALID",
                            "UNSUPPORTED_MEDIA_TYPE",
                            "VALIDATION_FAILED"
                          ],
                          "type": "string"
                        },
                        "context": {
                          "additionalProperties": {
                            "anyOf": [
                              {
                                "anyOf": [
                                  { "maxLength": 256, "type": "string" },
                                  {
                                    "maximum": 9007199254740991,
                                    "minimum": -9007199254740991,
                                    "type": "integer"
                                  },
                                  { "type": "boolean" },
                                  { "type": "null" }
                                ]
                              },
                              {
                                "items": {
                                  "anyOf": [
                                    { "maxLength": 256, "type": "string" },
                                    {
                                      "maximum": 9007199254740991,
                                      "minimum": -9007199254740991,
                                      "type": "integer"
                                    },
                                    { "type": "boolean" },
                                    { "type": "null" }
                                  ]
                                },
                                "maxItems": 50,
                                "type": "array"
                              }
                            ]
                          },
                          "propertyNames": {
                            "pattern": "^[A-Za-z][A-Za-z0-9]{0,63}$",
                            "type": "string"
                          },
                          "type": "object"
                        },
                        "fields": {
                          "additionalProperties": {
                            "items": {
                              "additionalProperties": false,
                              "properties": {
                                "code": {
                                  "enum": [
                                    "INVALID_FORMAT",
                                    "INVALID_TYPE",
                                    "INVALID_VALUE",
                                    "REQUIRED",
                                    "UNKNOWN_FIELD",
                                    "VALUE_TOO_LARGE",
                                    "VALUE_TOO_SMALL"
                                  ],
                                  "type": "string"
                                },
                                "message": { "maxLength": 256, "minLength": 1, "type": "string" }
                              },
                              "required": ["code", "message"],
                              "type": "object"
                            },
                            "maxItems": 10,
                            "minItems": 1,
                            "type": "array"
                          },
                          "propertyNames": { "maxLength": 256, "minLength": 1, "type": "string" },
                          "type": "object"
                        },
                        "message": { "maxLength": 256, "minLength": 1, "type": "string" },
                        "requestId": {
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                          "type": "string"
                        }
                      },
                      "required": ["code", "message", "requestId"],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": { "idempotentReplay": { "type": "boolean" } },
                      "required": ["idempotentReplay"],
                      "type": "object"
                    }
                  },
                  "required": ["error"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Get current session CSRF token",
        "tags": ["Account"]
      }
    },
...
  }
}